FEAT - configurado mensagens de retorno. - concluido função de sorteio.
