+++ amigoSecreto/Messages/__init__.py
@@ -1,0 +1,28 @@
+from enum import Enum
+
+class Mensagens(Enum):
+    EMAIL_TESTE = """
+    Olá, ####.
+    Está mensagem é de teste, confirme que recebeu, para sabermos que nenhum problema ocorreu!
+    """
+    MENSAGENS_AMIGO_SECRETO = [
+        """A baleia-azul é responsável pelo som mais alto do mundo. 
+        Sua voz pode ser ouvida até 800 km de distância. Ela só não ganha do seu amigo secreto #### gritando quando ganha algo que não gosta.""",
+        """Sem luzes para o Natal? Fiquei suave! A Maganize Luiza tem as melhores opções para você! Compre uma geladeira para #### no Natal e a encha. Você não pode perder a nossa oferta e perder a chance de ver o rosto se iluminar tanto, mais tanto, que nem vai precisar de pisca-pisca para a ceia.""",
+        """De lá para cá, de cá para lá, por onde vamos, carregamos conosco milhões de bactérias, e seu amigo secreto, que atende pelo nome de ####, também. ja parou para pensar que talvez sejamos apenas um meio de transporte para elas?""",
+        """O bico do ornitorrinco consegue detectar campos elétricos de outros seres vivos, fazendo com que ele tenha um sexto sentido para conseguir caçar e se mover. Ele consegue confiar completamente em seu bico, assim como seu amigo secreto #### confia que você não vai dar um par de meias para ele.""",
+        """Seu amigo secreto é um animal mamífero, bípede, que se distingue dos demais mamíferos pelo telencéfalo altamente desenvolvido e o polegar  opositor. Atende pelo nome de ####""",
+        """Se é pavê ou pacomê?
+        Pacomê eu não sei, mas pavê presente para ####, você vai ter que correr.""",
+        """
+        Olá, ####.
+
+        Te apresento seu amigo secreto Paulo Sergio.
+        De onde ele veio?
+        De trás do bananal
+        E o que ele merece?
+        Um presente bem legal.
+        """,
+        """Se ovo veio primeiro que a galinha, ou se a galinha veio primeiro que o ovo, eu não sei. Mas sei que #### de tal é seu amigo secreto, e isso me chocou. """,
+        """O aquecimento global ta aí, e todo mundo vai ser assado, então por que não contribuir para esse processo ser mais rápido?  É muito fácil para ajudar, de um presente para #### de tal e veja seu coração aquecer como o planeta, ele vai aquecer tanto que vai morrer de felicidade  :D"""
+    ]
