--- amigoSecreto/Messages/constantes.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-class Mensagens(Enum):
-    EMAIL_TESTE = """
-    Olá, ####.
-    Está mensagem é de teste, confirme que recebeu, para sabermos que nenhum problema ocorreu!
-    """
